fix(a11y): Add state totals to accessible table. Fixes #1692
