+++ src/components/pages/homepage/visualization-gallery/items/us-map/index.js
@@ -6,40 +6,8 @@ import { DateTime } from 'luxon'
 import metrics from './metrics'
 import Map from './map'
 import Grid from './grid'
+import Table from './table'
 import Title from '../../components/title'
-
-const Table = ({ states, metric, us }) => (
-  <div className="a11y-only">
-    <Link to={`#skip-state-list-${slugify(metric.title)}`}>
-      Skip the list of states
-    </Link>
-    <table>
-      <thead>
-        <tr>
-          <th>State</th>
-          <th>{metric.title}</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td>National total</td>
-          <td>{Math.round(us.value)}</td>
-        </tr>
-        {states.map(state => (
-          <tr>
-            <td>
-              <Link to={`/data/state/${state.childSlug.slug}`}>
-                {state.name}
-              </Link>
-            </td>
-            <td>{Math.round(state.value)}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-    <div id={`skip-state-list-${slugify(metric.title)}`} />
-  </div>
-)
 
 const USMap = ({ configuration, item }) => {
   const { metric } = configuration
